otel-collector/requirements.txt
pandas
numpy
matplotlib
scikit-learn
prophet
kafka-python
